Update: - Implemented deleting and editing both records and vehicles. - Added the feature to filter records by one vehicle at a time. - Fixed some bugs with theme changing. - Implemented a new window when loading app that check for differences between local and remote data and will notify the user as to which is more current and let the user choose for themselves. - Added a loading screen for when a user logs in and Firebase is getting data to compare to the local data.
(Previous)
Main update is added the feature to save vehicles and records to the local storage and get them displayed on their appropriate pages. Have not made a way to delete/update yet. Also added to where when you access the main pages, the local list of vehicles and records are uploaded to the Firebase Realtime Database for backup purposes. Have not implemented a way to restore data yet, but am planning on making it a feature, not a backend thing, that way teh local save is the master save point of the application.